requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
gitpython==3.1.44
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
iopath==0.1.10
ipykernel==6.29.5
ipython==9.3.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.4
jupyter-client==8.6.3
jupyter-core==5.8.1
lazy-loader==0.4
markupsafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
portalocker==3.1.1
prompt-toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
python-dateutil==2.9.0.post0
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@75ebeeaea0908c5527e7b1e305fbc7681382db47
pytz==2025.2
pyyaml==6.0.2
pyzmq==26.4.0
requests==2.32.4
scikit-image==0.25.2
scipy==1.15.3
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tifffile==2025.6.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing-extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
wandb==0.20.1
wcwidth==0.2.13
